+++ keyring/detectOther.ts
@@ -1,5 +1,5 @@
 
-import { packageInfo as utilInfo } from 'https://deno.land/x/polkadot@0.2.38/util/packageInfo.ts';
-import { packageInfo as cryptoInfo } from 'https://deno.land/x/polkadot@0.2.38/util-crypto/packageInfo.ts';
+import { packageInfo as utilInfo } from 'https://deno.land/x/polkadot/util/packageInfo.ts';
+import { packageInfo as cryptoInfo } from 'https://deno.land/x/polkadot/util-crypto/packageInfo.ts';
 
 export default [utilInfo, cryptoInfo];