+++ keyring/testing.ts
@@ -1,9 +1,9 @@
 
-import type { HexString } from 'https://deno.land/x/polkadot@0.2.38/util/types.ts';
-import type { KeypairType } from 'https://deno.land/x/polkadot@0.2.38/util-crypto/types.ts';
+import type { HexString } from 'https://deno.land/x/polkadot/util/types.ts';
+import type { KeypairType } from 'https://deno.land/x/polkadot/util-crypto/types.ts';
 import type { KeyringInstance, KeyringOptions } from './types.ts';
 
-import { hexToU8a } from 'https://deno.land/x/polkadot@0.2.38/util/mod.ts';
+import { hexToU8a } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { createPair } from './pair/index.ts';
 import { Keyring } from './keyring.ts';