+++ util-crypto/ed25519/sign.ts
@@ -1,11 +1,11 @@
 
-import type { HexString } from 'https://deno.land/x/polkadot@0.2.38/util/types.ts';
+import type { HexString } from 'https://deno.land/x/polkadot/util/types.ts';
 import type { Keypair } from '../types.ts';
 
 import { ed25519 } from 'https://esm.sh/@noble/curves@1.0.0/ed25519.js';
 
-import { hasBigInt, u8aToU8a } from 'https://deno.land/x/polkadot@0.2.38/util/mod.ts';
-import { ed25519Sign as wasmSign, isReady } from 'https://deno.land/x/polkadot@0.2.38/wasm-crypto/mod.ts';
+import { hasBigInt, u8aToU8a } from 'https://deno.land/x/polkadot/util/mod.ts';
+import { ed25519Sign as wasmSign, isReady } from 'https://deno.land/x/polkadot/wasm-crypto/mod.ts';
 
 /**
  * @name ed25519Sign
@@ -16,7 +16,7 @@ import { ed25519Sign as wasmSign, isReady } from 'https://deno.land/x/polkadot@0
  * <BR>
  *
  * ```javascript
- * import { ed25519Sign } from 'https://deno.land/x/polkadot@0.2.38/util-crypto/mod.ts';
+ * import { ed25519Sign } from 'https://deno.land/x/polkadot/util-crypto/mod.ts';
  *
  * ed25519Sign([...], [...]); // => [...]
  * ```