+++ util-crypto/networks.ts
@@ -1,2 +1,2 @@
 
-export { allNetworks, availableNetworks, selectableNetworks } from 'https://deno.land/x/polkadot@0.2.38/networks/mod.ts';
+export { allNetworks, availableNetworks, selectableNetworks } from 'https://deno.land/x/polkadot/networks/mod.ts';