+++ util/is/object.ts
@@ -13,7 +13,7 @@ interface ObjectIndexed {
  * <BR>
  *
  * ```javascript
- * import { isObject } from 'https://deno.land/x/polkadot@0.2.38/util/mod.ts';
+ * import { isObject } from 'https://deno.land/x/polkadot/util/mod.ts';
  *
  * isObject({}); // => true
  * isObject('something'); // => false