+++ x-ws/node.ts
@@ -1,7 +1,7 @@
 
 import ws from 'https://esm.sh/ws@8.13.0';
 
-import { extractGlobal } from 'https://deno.land/x/polkadot@0.2.38/x-global/mod.ts';
+import { extractGlobal } from 'https://deno.land/x/polkadot/x-global/mod.ts';
 
 export { packageInfo } from './packageInfo.ts';